# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIS-CUSTOMERS",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/regions": {
            "get": {
                "tags": [
                    "Regions"
                ],
                "summary": "Retrieve all regions",
                "description": "Get a list of all regions.",
                "operationId": "aa03b59c4f13cb98ebe59cdf5b96c3b1",
                "responses": {
                    "200": {
                        "description": "A list of regions"
                    }
                }
            },
            "post": {
                "tags": [
                    "Regions"
                ],
                "summary": "Create a new region",
                "description": "Create a new region with description and status.",
                "operationId": "f413ef316bea04b1c9ceb47d0677982b",
                "parameters": [
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Region description",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 90
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of the region",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "A",
                                "I",
                                "trash"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "maxLength": 90,
                                        "example": "New Region Description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I",
                                            "trash"
                                        ],
                                        "example": "A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Region created successfully"
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/regions/{id}": {
            "put": {
                "tags": [
                    "Regions"
                ],
                "summary": "Update data region",
                "description": "Update an existing region's description and status.",
                "operationId": "d7fcaf48a4432c9ff68f2a65cbb427cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Region ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Region description",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 90
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of the region",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "A",
                                "I",
                                "trash"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "maxLength": 90,
                                        "example": "Updated Region Description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "A",
                                            "I",
                                            "trash"
                                        ],
                                        "example": "I"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Region updated successfully"
                    },
                    "400": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Region not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Region not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Regions"
                ],
                "summary": "Delete region",
                "description": "Delete an existing region by ID.",
                "operationId": "567d7cae4e4b068d600e845a2d9519ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Region ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Region deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Region deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Region not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Region not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}